requirements.txt
# env
python-dotenv~=1.0.0
# base
pandas~=1.5.1
numpy==1.24.2
requests==2.28.2
scipy==1.11.3
# notebook
jupyterlab==3.5.0
# ml
scikit-learn==1.0.2
# graphifcs 
seaborn==0.11.0
matplotlib==3.3.3
# nlp
openai==0.25.0
tensorflow_text>=2.0.0rc0
nltk==3.6.1
# GCP
google-cloud-storage~=2.1.0